Collapse excess invasions (#2)
* Collapse excess invasions
