+++ src/components/panels/InvasionsPanel.vue
@@ -1,40 +1,23 @@
 <template>
   <HubPanelWrap :title="headertext">
     <b-list-group>
-      <b-list-group-item :style="styleObject" v-for="(invasion, index) in invasions" :key="invasion.id"
+      <b-list-group-item :style="styleObject" v-for="(invasion, index) in ongoing(invasions).splice(0,maxInvasions)" :key="invasion.id"
       v-bind:class="{
-        'list-group-item-borderless': index < (invasions.length - 1),
-        'list-group-item-borderbottom': index === (invasions.length - 1) }"
-        v-if="!invasion.completed">
-
-        <div :id="`${invasion.id}_info`" class="text-center">
-          <span><b>{{invasion.node}}</b></span>
-          <b-btn :id="`${invasion.id}_tooltip`" class="pull-right" size="sm" variant="secondary">?</b-btn>
-          <b-tooltip :target="`${invasion.id}_tooltip`" placement="top">
-            &nbsp;
-            <TimeBadge :starttime="invasion.activation" :counter="true" :interval="1000"/>
-          </b-tooltip>
-          <br/>
-          {{invasion.desc}} {{eta(invasion)}}
-        </div>
-        <b-row>
-          <b-col>
-            <div class="pull-left">
-              <b-badge tag="div" v-for="item in invasion.attackerReward.items" :key="item" :variant="getLabelColor(invasion.attackingFaction)">{{item}}</b-badge>
-              <b-badge tag="div" v-for="item in invasion.attackerReward.countedItems" :key="item.type" :variant="getLabelColor(invasion.attackingFaction)">{{countedItem(item)}}</b-badge>
-            </div>
-            <div class="pull-right">
-              <b-badge v-for="item in invasion.defenderReward.items" :key="item" :variant="getLabelColor(invasion.defendingFaction)">{{item}}</b-badge>
-              <b-badge v-for="item in invasion.defenderReward.countedItems" :key="item.type" :variant="getLabelColor(invasion.defendingFaction)">{{countedItem(item)}}</b-badge>
-            </div>
-          </b-col>
-        </b-row>
-        <b-row>
-          <b-progress :max="100" class="w-100 h-125">
-            <b-progress-bar :variant="getLabelColor(invasion.attackingFaction)" :value="invasion.completion"></b-progress-bar>
-            <b-progress-bar :variant="getLabelColor(invasion.defendingFaction)" :value="100 - invasion.completion"></b-progress-bar>
-          </b-progress>
-        </b-row>
+        'list-group-item-borderless': index < (ongoing(invasions).length - 1),
+        'list-group-item-borderbottom': index === (ongoing(invasions).length - 1) }">
+        <Invasion :invasion="invasion"></Invasion>
+      </b-list-group-item>
+      <b-list-group-item class="list-group-item-borderbottom" v-if="ongoing(invasions).length>2">
+        <Collapsible :headertext="`Show More (${ongoing(invasions).length-maxInvasions})`">
+          <b-list-group>
+            <b-list-group-item :style="styleObject" v-for="(invasion, index) in ongoing(invasions).splice(maxInvasions)" :key="invasion.id"
+              v-bind:class="{
+                'list-group-item-borderless': index < (ongoing(invasions).length - 1),
+                'list-group-item-borderbottom': index === (ongoing(invasions).length - 1) }">
+                <Invasion :invasion="invasion"></Invasion>
+            </b-list-group-item>
+          </b-list-group>
+        </Collapsible>
       </b-list-group-item>
       <NoDataItem v-if="invasions.length === 0" :text="headertext" />
     </b-list-group>
@@ -42,9 +25,10 @@
 </template>
 
 <script>
-  import TimeBadge from '@/components/TimeBadge.vue';
   import NoDataItem from '@/components/NoDataItem.vue';
   import HubPanelWrap from '@/components/HubPanelWrap';
+  import Collapsible from '@/components/Collapsible';
+  import Invasion from '@/components/Invasion';
 
   export default {
     name: 'InvasionsPanel',
@@ -53,6 +37,9 @@
       headertext() {
         return 'Invasions';
       },
+      maxInvasions() {
+        return 5
+      }
     },
     data () {
       return {
@@ -62,32 +49,19 @@
       };
     },
     methods: {
-      eta: function(invasion) {
-        return `(Ends in: ${invasion.eta.replace('-Infinityd', '??').replace('Infinityd', '??').replace(/\s\d\d?s/ig, '')})*`;
-      },
-      getLabelColor: function(faction) {
-        switch (faction) {
-          case 'Corpus':
-            return 'info';
-          case 'Grineer':
-            return 'danger';
-          case 'Infested':
-            return 'success';
-          case 'Corrupted':
-            return 'warning';
-          default:
-            return 'default';
+      ongoing: function(invasions) {
+        var ongoingInvasions = [];
+        for (var i = 0; i < invasions.length; i++) {
+          if (!invasions[i].completed) {
+            ongoingInvasions.push(invasions[i]);
+          }
         }
-      },
-      countedItem: function(item) {
-        if (item.count > 1) {
-          return `${item.count} ${item.type}`;
-        }
-        return item.type;
-      },
+        return ongoingInvasions;
+      }
     },
     components: {
-      TimeBadge,
+      Invasion,
+      Collapsible,
       NoDataItem,
       HubPanelWrap,
     }
